Moved onCreate and onOptionsItemSelected from MainActivity to HomeFragment
